requirements.txt
boto3
functions-framework
google-cloud-billing
google-cloud-secret-manager